+++ tests/testthat/test-parameters_table.R
@@ -4,14 +4,22 @@ skip_if_not_installed("insight")
 test_that("parameters_table 1", {
   x <- model_parameters(lm(Sepal.Length ~ Species, data = iris), standardize = "refit")
   tab <- insight::format_table(x)
-  expect_equal(colnames(tab), c("Parameter", "Coefficient", "SE", "95% CI", "t(147)", "p"))
+  expect_equal(
+    colnames(tab),
+    c("Parameter", "Coefficient", "SE", "95% CI", "t(147)", "p")
+  )
 })
 
 test_that("parameters_table 2", {
   skip_if_not_installed("lme4")
-  x <- model_parameters(lme4::lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris), effects = "fixed")
+  x <- model_parameters(
+    lme4::lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris),
+    effects = "fixed"
+  )
   tab <- insight::format_table(x)
-  expect_true(all(names(tab) == c("Parameter", "Coefficient", "SE", "95% CI", "t(146)", "p", "Effects")))
+  expect_true(all(
+    names(tab) == c("Parameter", "Coefficient", "SE", "95% CI", "t(146)", "p", "Effects")
+  ))
 })
 
 test_that("parameters_table 3", {